Fix a minor bug in the GraphQL formatter. (#30)
```
File /.../graphql_compiler/query_formatting/graphql_formatting.py", in leave_Directive
    arg = args.pop(arg_names.index(defined_arg))
IndexError: pop index out of range
```

This happened because the `args` and `arg_names` lists were expected to be parallel, but popping could happen in any order. The bug above is caused when the lists cease to be parallel, e.g. as follows:

```
********* before first pop, popping at index 0

['op_name: "="', 'value: ["$parameter"]']
['op_name', 'value']

********* before second pop, popping at index 1

['value: ["$parameter"]']
['op_name', 'value']

********* IndexError since there was no index
```