Fix fold_after_traverse bug
Previous query would give the wrong answer. Consider the following graph
structure, where all vertices are of type Animal and for vertices a and
b, a->b means a is the parent of b so there exists an edge
Animal_ParentOf from a to b.

Let the vertices be 1, 2, 3 with edges: 1->2, 1->3, 2->3. Then the
fold_after_traverse GraphQL query should give a result with 3 rows.

One row would have Animal___1 as 2 and the sibling_and_self_names_list as [2,
3]. One row would have Animal___1 as 3 and the sibling_and_self_names_list as
[3] (because the parent is 2). The third row would have Animal___1 as 3 again
but the sibling_and_self_names_list would be [2, 3] (because the parent in
this case is 2).

The old query would have merged the two list outputs for when Animal___1 is
3, yielding one row for when Animal___1 is 3. This is probably a quirk
of Cypher, since if we ran that query and also returned the parent node
(parent of Animal___1), then we do get the correct result. Cypher seems
to collect the two lists into one.

To get the correct results, the new query uses WITH to ensure that we do
get the three-row result.