Fix IR lowering for filtering and fold scopes (#464)
* Fix IR lowering for filtering and fold scopes

Previously, if we attempted to filter on or inside a fold scope, we'd
run into problems because the IR lowering would attempt to change the
LocalField within the filter to a FoldedContextField. However,
FoldedContextField's field type must be of type GraphQLList (because
it's for after the fold has happened).

In our case, we want to filter using the field belonging to the vertex
itself, pre-fold. We also can't use ContextField because that requires
its location be a Location type object, not a FoldScopeLocation object.

I've written a special class specifically to take care of this scenario.
Tests for these coming soon!

* Integrate new changes

* Update cypher-specific expression name and comment

* Lint

* Lint

* Rename class to more concise