Add fold tests for Cypher in test_compiler (#465)
* Add fold tests for Cypher in test_compiler

Right now in PR #437 I'm working on getting the fold functionality to
work. This PR contains the majority of compiler tests for fold for
Cypher, and can be reviewed as soon as PRs #459, #463, and #464 are
merged into master.

Four fold-related tests are still skipped for now: these tests include
optional in the same GraphQL query as fold and there may be some
potential edge cases I haven't fully fleshed out yet. Those tests are
already written; I've just left them out so that we can merge these
first and merge those tests along with the correct implementation for
OPTIONAL once that's taken care of.

* Fix test ordering

* Nits