Only start dockerized databases for tests, not for lint.
Linting time currently dominates test time. However, a large portion of the time taken is actually downloading the docker images for the various databases we test on -- which are not used at all in the lint checks. Not downloading the docker images in the lint configuration should even out the time taken by the various runs in the test matrix, and shorten the overall "time to approval" from Travis.