Update the Recurse block handler to use a "piggyback" strategy for inactive contexts.
The old implementation was O(N^2) for N levels of recursion. The new implementation is O(N) for N levels of recursion, while only using up to O(N) additional memory. The cost of N levels of recursion itself, purely for the depth of the stack, is also O(N), so this additional memory consumption is not significant.