Create codecov.yml allowing up to 0.03% coverage variance. (#806)
Our project coverage is already very high, to the point where a single line missing coverage will currently make the coverage check fail. This is somewhat unproductive, especially when such lines with changing coverage are frequently not in the code that was being modified, but a side-effect of some unrelated test change. We need to improve the depth of our coverage before it makes sense to be this strict, so that a single unrelated change can't cause a line to transition from covered to uncovered.

The 0.03% threshold allows a few lines (1 < N < 10) to transition from covered to uncovered, without reporting a negative patch coverage result.