Rename test_schema tests
To have them fit a consistent format: first, with specific details,
followed by whether it's a one-to-many rename or a suppression, and
finally if it's a one-to-many rename, whther or not it includes the
original field. If we have both a one-to-many rename and a suppression
in the same rename, the one-to-many rename comes first (e.g. in
test_one_to_many_rename_and_suppress).