Remove 1-many type tests
As described in the discussion of PR #828, we actually don't need to
support 1-many renaming for types, only for fields. That eliminates the
need for a number of tests that were previously here. This commit isn't
the end of the story though, since I plan to add in the 1-many argument
for rename_schema after this.