Add typing_extensions as a conditional dependency.
In compiler/ir_lowering_common/common.py we import `from typing_extensions import TypedDict`. However, `TypedDict` exists in the standard `typing` library as of Python 3.8: https://docs.python.org/3/library/typing.html#typing.TypedDict

Before that version, it has to be used from `typing_extensions`, but `typing_extensions` was not declared as a package dependency in such Python versions.

To resolve, adding a conditional import to the top-level typedefs.py file, and a conditional installation of typing_extensions for Python versions prior to 3.8.