Temporary placeholder commit for field renaming
If you're seeing this commit, it means I forgot to overwrite it and just
committed on top of it instead :)

Implemented: most of field renaming (and also got suppression for free,
it seems!) as well as some basic error checking. Next up: fix the test
cases that are raising errors when they shouldn't be, and fix the error
checking mechanisms.